+++ packages/terra-abstract-modal/src/AbstractModal.jsx
@@ -157,6 +157,11 @@ class AbstractModal extends React.Component {
       if (this.state.modalTrigger && this.state.modalTrigger.focus) {
         // Shift focus back to element that was last focused prior to opening the modal
         this.state.modalTrigger.focus();
+      } else {
+        // In some cases on IE, when the focus could not be restored on the element (SVG element, for instance)
+        // that was last focused prior to opening the modal, place the focus on the HTML body element to repro
+        // the behavior noticed on other major browsers.
+        document.querySelector('body').focus();
       }
     }, 0); // Allows inert processing to finish before shifting focus back
   }